64-bit fix from Joe Marcus Clarke <marcus@freebsd.org>
2004-06-12  Mark McLoughlin  <mark@skynet.ie>

        64-bit fix from Joe Marcus Clarke <marcus@freebsd.org>

        * backends/markup-tree.c: (parse_tree): the length
        returned from g_file_get_contents() is a gsize
        not an int.

+++ backends/markup-tree.c
@@ -3070,7 +3070,7 @@ parse_tree (MarkupDir  *root,
   ParseInfo info;
   char *filename;
   char *text;
-  int length;
+  gsize length;
 
   filename = markup_dir_build_path (root, TRUE, parse_subtree);
